commit20170417
The SMTP protocol includes a command to ask a server whether an address is valid. 

Usually VRFY is disabled to prevent spammers from finding legitimate email 
addresses,but if it is enabled you can ask the server about an address and
receive a status code indicating validity along with the user’s full name.